rename file and add help file for test
