+++ CMakeLists.txt
@@ -15,7 +15,8 @@ target_include_directories(${PROJECT_NAME} PRIVATE
 )
 
 target_sources(${PROJECT_NAME} PRIVATE
-        "${CMAKE_CURRENT_SOURCE_DIR}/sources/example.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/sources/bowlingGame.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/sources/helpForTest.cpp"
 )
 
 add_executable(${PROJECT_NAME}_executable "${CMAKE_CURRENT_SOURCE_DIR}/main.cpp")
