+++ includes/helpForTest.h
@@ -1,0 +1,10 @@
+#include <string>
+#include <iostream>
+#include <stdlib.h>
+#include <time.h>
+#include <vector>
+#include <numeric>
+
+using namespace std;
+
+bool checkSize(int count);
